Ajustados manejadores de señales
