+++ minishell.h
@@ -162,7 +162,7 @@ int				ft_setenv(t_abs_struct *base, t_process *p);
 void			ft_exit_minishell(t_abs_struct *base, int exit_code);
 int				ft_unset(t_abs_struct *base, t_process *p);
 int				ft_help(void);
-void			handle_sigint(int sig);
+void			signal_handler(int sig);
 int				ft_execute_command(t_abs_struct *base);
 int				ft_execute_ctrl_d(t_abs_struct *base);
 int				ft_init_minishell(t_abs_struct *base, char **envp);
