+++ srcs/minishell.c
@@ -93,6 +93,7 @@ int	main(int argc, char **argv, char **envp)
 		ft_exit_minishell(&g_base, 1);
 	while (minishell_ready)
 	{
+		signal(SIGINT, signal_handler);
 		ft_show_prompt(&g_base);
 		obtain_full_line();
 		execute_command_read(&g_base);
